added ternary operator to the login logout button
